pyproject.toml
[project]
name = "health-wellness-agent"
version = "0.1.0"
description = "A conversational AI-powered Health & Wellness Planner using Gemini API"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-generativeai==0.8.3",
    "pydantic==2.9.2",
    "streamlit==1.39.0",
    "python-dotenv==1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]